+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-# CHANGELOG
-
+* added access token authentication
 
 ## 1.5.0 (2023-02-22)