+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * added access token authentication
+* added CI test
 
 ## 1.5.0 (2023-02-22)