+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * added CI test
+
+## 1.5.1
+
 * added access token authentication
 
 ## 1.5.0 (2023-02-22)